+++ tests/Codeception/Acceptance/Basket/BasketsCest.php
@@ -24,6 +24,10 @@ final class BasketsCest extends BaseCest
 
     private const PASSWORD = 'useruser';
 
+    private const CUSTOMER_ID = 'e7af1c3b786fd02906ccd75698f4e6b9';
+
+    private const OTHER_CUSTOMER_ID = '245ad3b5380202966df6ff128e9eecaq';
+
     private const BASKET_ID = 'test_make_wishlist_private'; //owned by user@oxid-esales.com
 
     private const BASKET_ID_2 = '_test_basket_private'; //owned by otheruser@oxid-esales.com
@@ -31,6 +35,14 @@ final class BasketsCest extends BaseCest
     private const BASKET_ID_3 = '_test_wish_list_private'; //owned by otheruser@oxid-esales.com
 
     private const LAST_NAME = 'Muster';
+
+    public function _after(AcceptanceTester $I): void
+    {
+        $I->deleteFromDatabase('oxobject2group', ['OXID' => '_testrelationa']);
+        $I->deleteFromDatabase('oxobject2group', ['OXID' => '_testrelationb']);
+
+        parent::_after($I);
+    }
 
     public function testBasketsWithoutToken(AcceptanceTester $I): void
     {
@@ -86,23 +98,76 @@ final class BasketsCest extends BaseCest
         $I->assertSame(6, count($baskets));
     }
 
-    public function testBasketsCosts(AcceptanceTester $I): void
+    public function testPublicBasketsItemPriceForNotLoggedInUser(AcceptanceTester $I): void
     {
+        $I->wantToTest('that for a public basket no individual prices are displayed');
+
+        $this->assignSpecialPricesGroup($I);
+
+        $I->sendGQLQuery(
+            'query {
+                baskets(owner: "' . self::USERNAME . '") {
+                    id
+                    items {
+                        id
+                        product {
+                            price {
+                                price
+                            }
+                        }
+                    }
+                }
+            }'
+        );
+        $I->seeResponseIsJson();
+        $result = $I->grabJsonResponseAsArray();
+
+        $I->assertEquals('_test_basket_public', $result['data']['baskets'][1]['id']);
+        $I->assertEquals('_test_basket_item_1', $result['data']['baskets'][1]['items'][0]['id']);
+        $I->assertEquals(10, $result['data']['baskets'][1]['items'][0]['product']['price']['price']);
+
+        //check what the logged in basket owner will see
+        $I->login(self::USERNAME, self::PASSWORD);
+
+        $query = 'query {
+                    basket (basketId: "_test_basket_public") {
+                        items {
+                            product {
+                                id
+                                price {
+                                    price
+                                }
+                            }
+                        }
+                    }
+                }';
+
+        $I->sendGQLQuery($query);
+        $I->seeResponseIsJson();
+        $result = $I->grabJsonResponseAsArray();
+
+        $I->assertEquals(6.66, $result['data']['basket']['items'][0]['product']['price']['price']);
+    }
+
+    public function testPublicBasketsItemPriceForLoggedInSameUser(AcceptanceTester $I): void
+    {
+        $I->wantToTest('that for my own public baskets my group A prices are displayed');
+
+        $this->assignSpecialPricesGroup($I);
+
         $I->login(self::USERNAME, self::PASSWORD);
 
         $I->sendGQLQuery(
             'query {
                 baskets(owner: "' . self::USERNAME . '") {
                     id
-                    cost {
-                        productNet {
-                            price
+                    items {
+                        id
+                        product {
+                            price {
+                                price
+                            }
                         }
-                        payment {
-                            price
-                        }
-                        discount
-                        total
                     }
                 }
             }'
@@ -111,57 +176,41 @@ final class BasketsCest extends BaseCest
         $I->seeResponseIsJson();
         $result = $I->grabJsonResponseAsArray();
 
-        $I->assertSame([
-            [
-                'id'   => self::BASKET_ID,
-                'cost' => [
-                    'productNet' => [
-                        'price' => 0,
-                    ],
-                    'payment'    => [
-                        'price' => 0,
-                    ],
-                    'discount'   => 0,
-                    'total'      => 0,
-                ],
-            ], [
-                'id'   => '_test_basket_public',
-                'cost' => [
-                    'productNet' => [
-                        'price' => 8.4,
-                    ],
-                    'payment'    => [
-                        'price' => 7.5,
-                    ],
-                    'discount'   => 0,
-                    'total'      => 21.4,
-                ],
-            ], [
-                'id'   => '_test_voucher_public',
-                'cost' => [
-                    'productNet' => [
-                        'price' => 8.4,
-                    ],
-                    'payment'    => [
-                        'price' => 0,
-                    ],
-                    'discount'   => 0,
-                    'total'      => 13.9,
-                ],
-            ], [
-                'id'   => '_test_wish_list_public',
-                'cost' => [
-                    'productNet' => [
-                        'price' => 8.4,
-                    ],
-                    'payment'    => [
-                        'price' => 0,
-                    ],
-                    'discount'   => 0,
-                    'total'      => 13.9,
-                ],
-            ],
-        ], $result['data']['baskets']);
+        $I->assertEquals('_test_basket_public', $result['data']['baskets'][1]['id']);
+        $I->assertEquals('_test_basket_item_1', $result['data']['baskets'][1]['items'][0]['id']);
+        $I->assertEquals(6.66, $result['data']['baskets'][1]['items'][0]['product']['price']['price']);
+    }
+
+    public function testPublicBasketsItemPriceForLoggedInOtherUser(AcceptanceTester $I): void
+    {
+        $I->wantToTest('that for a public basket (not my own) my group b prices are displayed');
+
+        $this->assignSpecialPricesGroup($I);
+
+        $I->login(self::OTHER_USERNAME, self::PASSWORD);
+
+        $I->sendGQLQuery(
+            'query {
+                baskets(owner: "' . self::USERNAME . '") {
+                    id
+                    items {
+                        id
+                        product {
+                            price {
+                                price
+                            }
+                        }
+                    }
+                }
+            }'
+        );
+
+        $I->seeResponseIsJson();
+        $result = $I->grabJsonResponseAsArray();
+
+        $I->assertEquals('_test_basket_public', $result['data']['baskets'][1]['id']);
+        $I->assertEquals('_test_basket_item_1', $result['data']['baskets'][1]['items'][0]['id']);
+        $I->assertEquals(8.88, $result['data']['baskets'][1]['items'][0]['product']['price']['price']);
     }
 
     private function basketsQuery(AcceptanceTester $I, string $owner): array
@@ -179,7 +228,6 @@ final class BasketsCest extends BaseCest
                 }
                 id
                 title
-                public
                 creationDate
                 lastUpdateDate
             }
@@ -214,5 +262,29 @@ final class BasketsCest extends BaseCest
         $I->seeResponseIsJson();
 
         return $I->grabJsonResponseAsArray();
+    }
+
+    private function assignSpecialPricesGroup(AcceptanceTester $I): void
+    {
+        $I->haveInDatabase(
+            'oxobject2group',
+            [
+                'OXID'       => '_testrelationA',
+                'OXOBJECTID' => self::CUSTOMER_ID,
+                'OXGROUPSID' => 'oxidpricea',
+            ]
+        );
+
+        $I->haveInDatabase(
+            'oxobject2group',
+            [
+                'OXID'       => '_testrelationB',
+                'OXOBJECTID' => self::OTHER_CUSTOMER_ID,
+                'OXGROUPSID' => 'oxidpriceb',
+            ]
+        );
+
+        $I->updateInDatabase('oxarticles', ['OXPRICEA' => '6.66'], ['oxid' => '_test_product_for_basket']);
+        $I->updateInDatabase('oxarticles', ['OXPRICEB' => '8.88'], ['oxid' => '_test_product_for_basket']);
     }
 }